+++ country.html
@@ -55,6 +55,11 @@
       </nav>
     </header>
     <main class="uk-padding uk-flex-column">
+      <div
+        class="uk-width-1-1 uk-flex-center uk-flex search-form-wrapper"
+        id="searchInputWrapper"
+      ></div>
+
       <!-- Country Name Card -->
       <section
         class="card uk-flex-column uk-padding-small uk-margin text-white"
