+++ index.html
@@ -56,15 +56,10 @@
 
     <main class="uk-padding uk-flex-column">
       <!-- COUNTRY SEARCH -->
-      <div class="uk-width-1-1 uk-flex-center uk-flex search-form-wrapper">
-        <form id="searchForm" class="uk-width-medium">
-          <label class="uk-flex uk-flex-middle">
-            <i class="fas fa-search uk-margin-small-right text-white"></i>
-            <input type="text" placeholder="Country" class="uk-input" />
-            <div id="searchFormDropdown"></div>
-          </label>
-        </form>
-      </div>
+      <div
+        class="uk-width-1-1 uk-flex-center uk-flex search-form-wrapper"
+        id="searchInputWrapper"
+      ></div>
 
       <section
         id="worldData"
